Refactor #8 and #9
This avoids the use of CPP where possible; the last remaining CPP
use-site could in principle be avoided too by using cabal-level
conditional compilation but for now we've just reduced the source-code
exposed to the CPP phase to a bare minimum.

While at it, also fixup the package description and upgrade to
cabal-version:2.0 as there's little reason not to in 2021.